refactor(http): Decouple Router from response sending
Removes the responsibility of sending the HTTP response from the Router class.

The dispatch and invokeHandler methods no longer call ->send() on the response object. The final ->send() call is now handled at the end of the application lifecycle in index.php. This change improves separation of concerns and makes the Router fully unit-testable.